+++ CHANGELOG.md
@@ -5,7 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## [0.6.0](https://github.com/returntocorp/semgrep/releases/tag/v0.6.0) - 2020-05-05
 
 ### Added
+- `--exclude`, `--include`, `--exclude-dir` flags for targeting specific paths with command line options.
+  The behavior of these flags mimics `grep`'s behavior.
+
+### Added
 - The `-j/--jobs` flag for specifying number of subprocesses to use to run checks in parallel.
+
+### Added
 - expression statements will now match by default also return statements
   ```
   foo();